refactor(git): extract sync/history commands into git_sync.py; git_cmd.py is now a thin orchestrator
